feat, chore, refactor: Addition of functionality to the four arithmetic buttons and little change in the README.md file.
feat --> The 'sum', 'substraction', 'division' and 'multiplication'
buttons has a little new functionality. The topScreen in the calculator
begins to work, but it's still incomplete.

refactor --> Refactor made in the calculatorFunctionality.js file, in
the 'buttonAction', 'processValue' and 'bottomScreenPrint' functions, in
the required parameters. With a little change in the 'setUp' function too.

chore --> Implementation of the actual state of the project in the README.md file